+++ CHANGELOG.md
@@ -111,6 +111,7 @@
 91. Fixed Logger Flushing.
 92. Fixed Issue with Console Log Coloring.
 93. Redirected 0Harmony.dll References to MelonLoader.dll.
+94. Added Current Culture Fix to ensure Culture is always Invariant.
 
 ---